Finished writing the code for interrupting and determining which of the spell buttons is pressed and controlling the corresponding LEDs. Also implemented the basic logic needed to transition from state to state and determine valid inputs.
Added timer 1 to the list of configured modules. This is used to determine how long the detonate button has been held down to determine arming mode. Laid the groundwork for communicating the selection via Bluetooth from the remote to the mine.

Also adjusted the configuration of the UART module to meet the specifications given in the Bluetooth data sheet.

To do:  Rename remote_common to remote_util to more clearly indicate its purpose.
	Lay groundwork to communicate via Bluetooth which of the spells has been selected.
	Lots of other stuff.